# lnwire: fix heap escapes to reduce gc pressure