--- a/lnwire/closing_signed.go
+++ b/lnwire/closing_signed.go
@@ -1,6 +1,7 @@
 package lnwire
 
 import (
+	"bytes"
 	"io"
 
 	"github.com/btcsuite/btcutil"
@@ -63,7 +64,7 @@ func (c *ClosingSigned) Decode(r io.Reader, pver uint32) error {
 // observing the protocol version specified.
 //
 // This is part of the lnwire.Message interface.
-func (c *ClosingSigned) Encode(w io.Writer, pver uint32) error {
+func (c *ClosingSigned) Encode(w *bytes.Buffer, pver uint32) error {
 	return WriteElements(
 		w, c.ChannelID, c.FeeSatoshis, c.Signature, c.ExtraData,
 	)