--- a/lnwire/extra_bytes.go
+++ b/lnwire/extra_bytes.go
@@ -18,7 +18,7 @@ type ExtraOpaqueData []byte
 // Encode attempts to encode the raw extra bytes into the passed io.Writer.
 func (e *ExtraOpaqueData) Encode(w *bytes.Buffer) error {
 	eBytes := []byte((*e)[:])
-	if err := WriteElements(w, eBytes); err != nil {
+	if err := WriteBytes(w, eBytes); err != nil {
 		return err
 	}