This PR is meant to fix #3004.

The overall optimization is done in three steps,
- change `WriteMessage` to use write buffer.
- change `Encode` and `WriteElement` to use write buffer.
- replace interface types with concrete types inside `Encode`.

The result,
```bash
name                                    old time/op    new time/op    delta
WriteMessage/Init-4                       3.62µs ±25%    2.29µs ±11%   -36.70%  (p=0.000 n=10+10)
WriteMessage/Error-4                      1.03µs ±10%    0.37µs ± 9%   -63.94%  (p=0.000 n=10+10)
WriteMessage/Ping-4                        977ns ±17%     352ns ± 7%   -63.98%  (p=0.000 n=10+10)
WriteMessage/Pong-4                        834ns ± 1%     336ns ± 3%   -59.72%  (p=0.000 n=8+10)
WriteMessage/MsgOpenChannel-4             3.54µs ± 0%    1.65µs ± 3%   -53.30%  (p=0.000 n=7+10)
WriteMessage/MsgAcceptChannel-4           3.03µs ± 4%    1.65µs ± 2%   -45.48%  (p=0.000 n=10+9)
WriteMessage/MsgFundingCreated-4          1.34µs ± 0%    0.36µs ± 1%   -73.36%  (p=0.000 n=6+10)
WriteMessage/MsgFundingSigned-4           1.26µs ± 3%    0.34µs ± 1%   -73.00%  (p=0.000 n=9+10)
WriteMessage/FundingLocked-4              1.21µs ± 1%    0.45µs ± 1%   -62.51%  (p=0.000 n=8+9)
WriteMessage/Shutdown-4                   1.01µs ± 1%    0.34µs ± 3%   -66.18%  (p=0.000 n=9+10)
WriteMessage/ClosingSigned-4              1.17µs ± 2%    0.37µs ± 3%   -68.86%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4              2.21µs ± 1%    0.38µs ± 1%   -82.71%  (p=0.000 n=8+8)
WriteMessage/UpdateFulfillHTLC-4          1.14µs ± 2%    0.33µs ± 0%   -70.66%  (p=0.000 n=9+9)
WriteMessage/UpdateFailHTLC-4             1.01µs ± 1%    0.35µs ± 2%   -65.60%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                  11.8µs ± 2%     1.4µs ± 1%   -88.00%  (p=0.000 n=9+9)
WriteMessage/RevokeAndAck-4               1.26µs ± 1%    0.47µs ± 2%   -63.02%  (p=0.000 n=9+9)
WriteMessage/UpdateFee-4                   960ns ± 1%     329ns ± 1%   -65.69%  (p=0.000 n=8+10)
WriteMessage/UpdateFailMalformedHTLC-4    1.19µs ± 1%    0.35µs ± 2%   -70.95%  (p=0.000 n=9+10)
WriteMessage/ChannelReestablish-4         1.34µs ± 1%    0.48µs ± 1%   -63.74%  (p=0.000 n=8+9)
WriteMessage/ChannelAnnouncement-4        3.01µs ± 1%    1.23µs ± 2%   -59.27%  (p=0.000 n=10+9)
WriteMessage/NodeAnnouncement-4           3.33µs ± 3%    2.99µs ± 1%   -10.37%  (p=0.000 n=9+9)
WriteMessage/ChannelUpdate-4              1.42µs ± 1%    0.43µs ± 9%   -69.71%  (p=0.000 n=9+9)
WriteMessage/AnnounceSignatures-4         1.45µs ± 7%    0.40µs ±12%   -72.19%  (p=0.000 n=9+10)
WriteMessage/QueryShortChanIDs-4           161µs ± 7%      75µs ± 2%   -53.30%  (p=0.000 n=9+10)
WriteMessage/ReplyShortChanIDsEnd-4        917ns ± 1%     326ns ± 1%   -64.41%  (p=0.000 n=9+9)
WriteMessage/QueryChannelRange-4           975ns ± 1%     337ns ± 1%   -65.49%  (p=0.000 n=8+9)
WriteMessage/ReplyChannelRange-4           145µs ± 1%      74µs ± 1%   -48.87%  (p=0.000 n=9+10)
WriteMessage/GossipTimestampRange-4       1.01µs ±10%    0.34µs ± 1%   -66.38%  (p=0.000 n=10+9)
WriteMessage/QueryShortChanIDs#01-4        712µs ±23%     408µs ± 3%   -42.65%  (p=0.000 n=10+10)
WriteMessage/ReplyChannelRange#01-4        681µs ± 3%     406µs ± 2%   -40.40%  (p=0.000 n=9+9)

name                                    old alloc/op   new alloc/op   delta
WriteMessage/Init-4                       1.33kB ± 0%    0.02kB ± 0%   -98.80%  (p=0.000 n=10+10)
WriteMessage/Error-4                      1.36kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/Ping-4                       1.30kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/Pong-4                       1.29kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/MsgOpenChannel-4             5.84kB ± 0%    1.75kB ± 0%   -70.00%  (p=0.000 n=10+10)
WriteMessage/MsgAcceptChannel-4           4.57kB ± 0%    1.75kB ± 0%   -61.65%  (p=0.000 n=9+10)
WriteMessage/MsgFundingCreated-4          1.82kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/MsgFundingSigned-4           1.95kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/FundingLocked-4              1.94kB ± 0%    0.08kB ± 0%   -95.87%  (p=0.000 n=10+10)
WriteMessage/Shutdown-4                   1.40kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/ClosingSigned-4              1.98kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4              7.14kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFulfillHTLC-4          1.84kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFailHTLC-4             1.40kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                  29.2kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/RevokeAndAck-4               1.96kB ± 0%    0.08kB ± 0%   -95.92%  (p=0.000 n=10+10)
WriteMessage/UpdateFee-4                  1.39kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFailMalformedHTLC-4    1.85kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/ChannelReestablish-4         1.98kB ± 0%    0.08kB ± 0%   -95.97%  (p=0.000 n=10+10)
WriteMessage/ChannelAnnouncement-4        3.80kB ± 0%    0.01kB ± 0%   -99.79%  (p=0.000 n=10+10)
WriteMessage/NodeAnnouncement-4           3.16kB ± 0%    0.09kB ± 0%   -97.22%  (p=0.000 n=10+10)
WriteMessage/ChannelUpdate-4              1.95kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/AnnounceSignatures-4         2.11kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs-4          64.2kB ± 0%     0.1kB ± 1%   -99.86%  (p=0.000 n=10+10)
WriteMessage/ReplyShortChanIDsEnd-4       1.34kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryChannelRange-4          1.36kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/ReplyChannelRange-4          64.2kB ± 0%     0.1kB ± 0%   -99.86%  (p=0.000 n=10+8)
WriteMessage/GossipTimestampRange-4       1.36kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs#01-4        876kB ± 0%     841kB ± 0%    -3.95%  (p=0.000 n=7+10)
WriteMessage/ReplyChannelRange#01-4        876kB ± 0%     841kB ± 0%    -3.95%  (p=0.000 n=9+10)

name                                    old allocs/op  new allocs/op  delta
WriteMessage/Init-4                         10.0 ± 0%       4.0 ± 0%   -60.00%  (p=0.000 n=10+10)
WriteMessage/Error-4                        8.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
WriteMessage/Ping-4                         8.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
WriteMessage/Pong-4                         6.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
WriteMessage/MsgOpenChannel-4               55.0 ± 0%      19.0 ± 0%   -65.45%  (p=0.000 n=10+10)
WriteMessage/MsgAcceptChannel-4             48.0 ± 0%      19.0 ± 0%   -60.42%  (p=0.000 n=10+10)
WriteMessage/MsgFundingCreated-4            13.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/MsgFundingSigned-4             11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/FundingLocked-4                12.0 ± 0%       2.0 ± 0%   -83.33%  (p=0.000 n=10+10)
WriteMessage/Shutdown-4                     10.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/ClosingSigned-4                13.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4                18.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFulfillHTLC-4            12.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFailHTLC-4               11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                     217 ± 0%         0       -100.00%  (p=0.000 n=10+10)
WriteMessage/RevokeAndAck-4                 13.0 ± 0%       2.0 ± 0%   -84.62%  (p=0.000 n=10+10)
WriteMessage/UpdateFee-4                    10.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFailMalformedHTLC-4      15.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/ChannelReestablish-4           17.0 ± 0%       2.0 ± 0%   -88.24%  (p=0.000 n=10+10)
WriteMessage/ChannelAnnouncement-4          31.0 ± 0%       2.0 ± 0%   -93.55%  (p=0.000 n=10+10)
WriteMessage/NodeAnnouncement-4             39.0 ± 0%       4.0 ± 0%   -89.74%  (p=0.000 n=10+10)
WriteMessage/ChannelUpdate-4                27.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/AnnounceSignatures-4           17.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs-4           4.02k ± 0%     0.00k ± 0%   -99.93%  (p=0.000 n=10+10)
WriteMessage/ReplyShortChanIDsEnd-4         8.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryChannelRange-4            11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/ReplyChannelRange-4           4.02k ± 0%     0.00k ± 0%   -99.93%  (p=0.000 n=10+10)
WriteMessage/GossipTimestampRange-4         11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs#01-4        4.04k ± 0%     0.03k ± 0%   -99.18%  (p=0.000 n=10+10)
WriteMessage/ReplyChannelRange#01-4        4.04k ± 0%     0.03k ± 0%   -99.18%  (p=0.000 n=10+10)
```

Top heap allocs,
```bash
(pprof) top
Showing nodes accounting for 79.53GB, 98.26% of 80.94GB total
Dropped 61 nodes (cum <= 0.40GB)
Showing top 10 nodes out of 39
      flat  flat%   sum%        cum   cum%
   37.77GB 46.66% 46.66%    45.84GB 56.64%  compress/flate.NewWriter
   15.61GB 19.28% 65.94%    17.82GB 22.02%  github.com/lightningnetwork/lnd/lnwire.packShutdownScript
    8.87GB 10.96% 76.91%    14.73GB 18.20%  github.com/btcsuite/btcd/btcec.(*PublicKey).SerializeCompressed
    7.84GB  9.69% 86.60%     7.84GB  9.69%  compress/flate.(*compressor).initDeflate (inline)
    5.86GB  7.24% 93.83%     5.86GB  7.24%  math/big.(*Int).Bytes (inline)
    1.57GB  1.94% 95.78%     1.57GB  1.94%  bytes.makeSlice
    0.88GB  1.08% 96.86%     2.45GB  3.03%  bytes.(*Buffer).grow
    0.66GB  0.81% 97.67%     1.98GB  2.45%  github.com/lightningnetwork/lnd/lnwire.(*ExtraOpaqueData).PackRecords
    0.46GB  0.57% 98.24%     0.46GB  0.57%  github.com/lightningnetwork/lnd/tlv.NewStream
    0.02GB 0.021% 98.26%    47.87GB 59.14%  github.com/lightningnetwork/lnd/lnwire.encodeShortChanIDs
```

### Build the benchmark
This first step is to build the benchmark for our optimization. Run the following commands to build the profiles. Notice that we need to run the bench test multiple times (I've chosen 10) to gather meaningful statistics for `benchstat` to analyze.

```bash
cd lnwire

gco 0c409c59

go test -run=^$ -bench=Write -count=10 -benchmem -memprofile=write-mem-base.prof -cpuprofile=write-cpu-base.prof -o lnwire-base.test | tee write-bench-base.prof
```

Now we can check the heap allocs using,

```bash
go tool pprof --sample_index=alloc_space lnwire-base.test write-mem-base.prof
```

And checking the heap escapes using `go build -gcflags "-m -m"`.


### First optimiztion - Using write buffer in `WriteMessage`

Run the following commands to gather stats.
```bash
gco e2a51cf2

go test -run=^$ -bench=Write -count=10 -benchmem -memprofile=write-mem-1.prof -cpuprofile=write-cpu-1.prof -o lnwire-1.test | tee write-bench-1.prof
```

See the improvement,

```bash
benchstat write-bench-base.prof write-bench-1.prof
```

```bash
name                                    old time/op    new time/op    delta
WriteMessage/Init-4                       3.62µs ±25%    2.36µs ± 5%  -34.81%  (p=0.000 n=10+8)
WriteMessage/Error-4                      1.03µs ±10%    0.53µs ±16%  -48.73%  (p=0.000 n=10+9)
WriteMessage/Ping-4                        977ns ±17%     709ns ±46%  -27.37%  (p=0.001 n=10+10)
WriteMessage/Pong-4                        834ns ± 1%     483ns ± 5%  -42.13%  (p=0.000 n=8+8)
WriteMessage/MsgOpenChannel-4             3.54µs ± 0%    3.05µs ±16%  -13.78%  (p=0.000 n=7+8)
WriteMessage/MsgAcceptChannel-4           3.03µs ± 4%    2.41µs ±13%  -20.72%  (p=0.000 n=10+10)
WriteMessage/MsgFundingCreated-4          1.34µs ± 0%    0.74µs ±12%  -44.88%  (p=0.000 n=6+9)
WriteMessage/MsgFundingSigned-4           1.26µs ± 3%    0.72µs ±24%  -42.60%  (p=0.000 n=9+10)
WriteMessage/FundingLocked-4              1.21µs ± 1%    0.82µs ±19%  -32.25%  (p=0.000 n=8+9)
WriteMessage/Shutdown-4                   1.01µs ± 1%    0.62µs ±10%  -38.19%  (p=0.000 n=9+10)
WriteMessage/ClosingSigned-4              1.17µs ± 2%    0.69µs ± 7%  -41.31%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4              2.21µs ± 1%    0.75µs ± 1%  -66.32%  (p=0.000 n=8+8)
WriteMessage/UpdateFulfillHTLC-4          1.14µs ± 2%    0.63µs ± 5%  -44.89%  (p=0.000 n=9+10)
WriteMessage/UpdateFailHTLC-4             1.01µs ± 1%    0.64µs ± 9%  -36.30%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                  11.8µs ± 2%     8.9µs ± 9%  -24.86%  (p=0.000 n=9+10)
WriteMessage/RevokeAndAck-4               1.26µs ± 1%    0.80µs ± 9%  -36.21%  (p=0.000 n=9+10)
WriteMessage/UpdateFee-4                   960ns ± 1%     626ns ±17%  -34.80%  (p=0.000 n=8+10)
WriteMessage/UpdateFailMalformedHTLC-4    1.19µs ± 1%    0.73µs ± 7%  -38.80%  (p=0.000 n=9+8)
WriteMessage/ChannelReestablish-4         1.34µs ± 1%    0.92µs ± 8%  -31.42%  (p=0.000 n=8+8)
WriteMessage/ChannelAnnouncement-4        3.01µs ± 1%    2.28µs ± 7%  -24.15%  (p=0.000 n=10+10)
WriteMessage/NodeAnnouncement-4           3.33µs ± 3%    2.53µs ± 0%  -24.01%  (p=0.000 n=9+8)
WriteMessage/ChannelUpdate-4              1.42µs ± 1%    0.95µs ± 6%  -32.92%  (p=0.000 n=9+10)
WriteMessage/AnnounceSignatures-4         1.45µs ± 7%    0.93µs ±11%  -35.66%  (p=0.000 n=9+8)
WriteMessage/QueryShortChanIDs-4           161µs ± 7%     155µs ±12%     ~     (p=0.400 n=9+10)
WriteMessage/ReplyShortChanIDsEnd-4        917ns ± 1%     630ns ±21%  -31.24%  (p=0.000 n=9+10)
WriteMessage/QueryChannelRange-4           975ns ± 1%     655ns ±11%  -32.82%  (p=0.000 n=8+9)
WriteMessage/ReplyChannelRange-4           145µs ± 1%     146µs ± 9%     ~     (p=1.000 n=9+9)
WriteMessage/GossipTimestampRange-4       1.01µs ±10%    0.58µs ± 4%  -42.41%  (p=0.000 n=10+8)
WriteMessage/QueryShortChanIDs#01-4        712µs ±23%     762µs ± 4%     ~     (p=0.122 n=10+8)
WriteMessage/ReplyChannelRange#01-4        681µs ± 3%     600µs ±10%  -11.81%  (p=0.000 n=9+10)

name                                    old alloc/op   new alloc/op   delta
WriteMessage/Init-4                       1.33kB ± 0%    0.06kB ± 0%  -95.18%  (p=0.000 n=10+10)
WriteMessage/Error-4                      1.36kB ± 0%    0.09kB ± 0%  -93.36%  (p=0.000 n=10+10)
WriteMessage/Ping-4                       1.30kB ± 0%    0.03kB ± 0%  -97.69%  (p=0.000 n=10+10)
WriteMessage/Pong-4                       1.29kB ± 0%    0.03kB ± 0%  -97.99%  (p=0.000 n=10+10)
WriteMessage/MsgOpenChannel-4             5.84kB ± 0%    2.26kB ± 0%  -61.37%  (p=0.000 n=10+10)
WriteMessage/MsgAcceptChannel-4           4.57kB ± 0%    2.19kB ± 0%  -52.01%  (p=0.000 n=9+10)
WriteMessage/MsgFundingCreated-4          1.82kB ± 0%    0.28kB ± 0%  -84.51%  (p=0.000 n=10+10)
WriteMessage/MsgFundingSigned-4           1.95kB ± 0%    0.24kB ± 0%  -87.72%  (p=0.000 n=10+10)
WriteMessage/FundingLocked-4              1.94kB ± 0%    0.24kB ± 0%  -87.62%  (p=0.000 n=10+10)
WriteMessage/Shutdown-4                   1.40kB ± 0%    0.14kB ± 0%  -90.17%  (p=0.000 n=10+10)
WriteMessage/ClosingSigned-4              1.98kB ± 0%    0.26kB ± 0%  -87.04%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4              7.14kB ± 0%    0.21kB ± 0%  -97.09%  (p=0.000 n=10+10)
WriteMessage/UpdateFulfillHTLC-4          1.84kB ± 0%    0.15kB ± 0%  -91.74%  (p=0.000 n=10+10)
WriteMessage/UpdateFailHTLC-4             1.40kB ± 0%    0.14kB ± 0%  -90.29%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                  29.2kB ± 0%    13.1kB ± 0%  -55.22%  (p=0.000 n=10+10)
WriteMessage/RevokeAndAck-4               1.96kB ± 0%    0.26kB ± 0%  -86.54%  (p=0.000 n=10+10)
WriteMessage/UpdateFee-4                  1.39kB ± 0%    0.12kB ± 0%  -91.35%  (p=0.000 n=10+10)
WriteMessage/UpdateFailMalformedHTLC-4    1.85kB ± 0%    0.17kB ± 0%  -90.91%  (p=0.000 n=10+10)
WriteMessage/ChannelReestablish-4         1.98kB ± 0%    0.30kB ± 0%  -85.08%  (p=0.000 n=10+10)
WriteMessage/ChannelAnnouncement-4        3.80kB ± 0%    1.00kB ± 0%  -73.68%  (p=0.000 n=10+10)
WriteMessage/NodeAnnouncement-4           3.16kB ± 0%    0.79kB ± 0%  -75.06%  (p=0.000 n=10+10)
WriteMessage/ChannelUpdate-4              1.95kB ± 0%    0.27kB ± 0%  -86.07%  (p=0.000 n=10+10)
WriteMessage/AnnounceSignatures-4         2.11kB ± 0%    0.39kB ± 0%  -81.29%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs-4          64.2kB ± 0%    26.8kB ± 0%  -58.21%  (p=0.000 n=10+10)
WriteMessage/ReplyShortChanIDsEnd-4       1.34kB ± 0%    0.07kB ± 0%  -94.55%  (p=0.000 n=10+10)
WriteMessage/QueryChannelRange-4          1.36kB ± 0%    0.09kB ± 0%  -93.51%  (p=0.000 n=10+10)
WriteMessage/ReplyChannelRange-4          64.2kB ± 0%    26.8kB ± 0%  -58.20%  (p=0.000 n=10+10)
WriteMessage/GossipTimestampRange-4       1.36kB ± 0%    0.09kB ± 0%  -93.51%  (p=0.000 n=10+10)
WriteMessage/QueryShortChanIDs#01-4        876kB ± 0%     849kB ± 0%   -3.05%  (p=0.000 n=7+10)
WriteMessage/ReplyChannelRange#01-4        876kB ± 0%     849kB ± 0%   -3.05%  (p=0.000 n=9+10)

name                                    old allocs/op  new allocs/op  delta
WriteMessage/Init-4                         10.0 ± 0%       6.0 ± 0%  -40.00%  (p=0.000 n=10+10)
WriteMessage/Error-4                        8.00 ± 0%      4.00 ± 0%  -50.00%  (p=0.000 n=10+10)
WriteMessage/Ping-4                         8.00 ± 0%      4.00 ± 0%  -50.00%  (p=0.000 n=10+10)
WriteMessage/Pong-4                         6.00 ± 0%      2.00 ± 0%  -66.67%  (p=0.000 n=10+10)
WriteMessage/MsgOpenChannel-4               55.0 ± 0%      48.0 ± 0%  -12.73%  (p=0.000 n=10+10)
WriteMessage/MsgAcceptChannel-4             48.0 ± 0%      42.0 ± 0%  -12.50%  (p=0.000 n=10+10)
WriteMessage/MsgFundingCreated-4            13.0 ± 0%       8.0 ± 0%  -38.46%  (p=0.000 n=10+10)
WriteMessage/MsgFundingSigned-4             11.0 ± 0%       6.0 ± 0%  -45.45%  (p=0.000 n=10+10)
WriteMessage/FundingLocked-4                12.0 ± 0%       7.0 ± 0%  -41.67%  (p=0.000 n=10+10)
WriteMessage/Shutdown-4                     10.0 ± 0%       6.0 ± 0%  -40.00%  (p=0.000 n=10+10)
WriteMessage/ClosingSigned-4                13.0 ± 0%       8.0 ± 0%  -38.46%  (p=0.000 n=10+10)
WriteMessage/UpdateAddHTLC-4                18.0 ± 0%      12.0 ± 0%  -33.33%  (p=0.000 n=10+10)
WriteMessage/UpdateFulfillHTLC-4            12.0 ± 0%       7.0 ± 0%  -41.67%  (p=0.000 n=10+10)
WriteMessage/UpdateFailHTLC-4               11.0 ± 0%       7.0 ± 0%  -36.36%  (p=0.000 n=10+10)
WriteMessage/CommitSig-4                     217 ± 0%       208 ± 0%   -4.15%  (p=0.000 n=10+10)
WriteMessage/RevokeAndAck-4                 13.0 ± 0%       8.0 ± 0%  -38.46%  (p=0.000 n=10+10)
WriteMessage/UpdateFee-4                    10.0 ± 0%       6.0 ± 0%  -40.00%  (p=0.000 n=10+10)
WriteMessage/UpdateFailMalformedHTLC-4      15.0 ± 0%      10.0 ± 0%  -33.33%  (p=0.000 n=10+10)
WriteMessage/ChannelReestablish-4           17.0 ± 0%      12.0 ± 0%  -29.41%  (p=0.000 n=10+10)
WriteMessage/ChannelAnnouncement-4          31.0 ± 0%      25.0 ± 0%  -19.35%  (p=0.000 n=10+10)
WriteMessage/NodeAnnouncement-4             39.0 ± 0%      33.0 ± 0%  -15.38%  (p=0.000 n=10+10)
WriteMessage/ChannelUpdate-4                27.0 ± 0%      22.0 ± 0%  -18.52%  (p=0.000 n=10+10)
WriteMessage/AnnounceSignatures-4           17.0 ± 0%      12.0 ± 0%  -29.41%  (p=0.000 n=10+10)
…