@Crypt-iQ 2021-01-27
Nice work! Drive-thru comment: commits should compile, and perhaps the diff could be broken up into smaller PRs to make review easier.