@Crypt-iQ 2021-01-29
I think the `WriteMessage` change from `io.Writer` to `*bytes.Buffer` can be a separate commit from `WriteElements` change from `io.Writer` to `*bytes.Buffer`.  One gain from `WriteMessage` is the removal of `var bw bytes.Buffer`.  Splitting up the commit there can help see incrementally the gains realized.