@Crypt-iQ 2021-06-22
> `make lint` gave me this weird error, does anybody know what happened?
> 
> ```go
> lnwire/writer.go:50:17: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteBytes(buf *bytes.Buffer, b []byte) error {
>                 ^
> lnwire/writer.go:56:17: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteUint8(buf *bytes.Buffer, n uint8) error {
>                 ^
> lnwire/writer.go:63:18: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteUint16(buf *bytes.Buffer, n uint16) error {
> ```

No idea, that is very weird.