@carlaKC 2021-06-23
> `make lint` gave me this weird error, does anybody know what happened?
> 
> ```go
> lnwire/writer.go:50:17: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteBytes(buf *bytes.Buffer, b []byte) error {
>                 ^
> lnwire/writer.go:56:17: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteUint8(buf *bytes.Buffer, n uint8) error {
>                 ^
> lnwire/writer.go:63:18: `buf` can be `github.com/miekg/dns.Writer` (interfacer)
> func WriteUint16(buf *bytes.Buffer, n uint16) error {
> ```

If you only use the methods of `bytes.Buffer` that are implemented by `io.Writer` then the linter wants you to just set the parameter to the interface since that's all you need. 

another example, this would fail for the linter because you could just use the `Stringer` interface here: 
```
func something(hash lntypes.Hash)string{
    return hash.String()
}
```

Since we intentionally want to replace the interface here, you can just add `//nolint: interfacer` and perhaps a comment indicating why we don't use the interface.