@yyforyongyu 2021-06-23
> Since we intentionally want to replace the interface here, you can just add `//nolint: interfacer` and perhaps a comment indicating why we don't use the interface.

Cool, good to know, thanks! To me it's weird that the linter is asking me to use the package `github.com/miekg/dns.Writer`, which I don't think it's in the standard library.