@carlaKC 2021-06-23
> Cool, good to know, thanks! To me it's weird that the linter is asking me to use the package `github.com/miekg/dns.Writer`, which I don't think it's in the standard library.

Ah yeah, probably just happens to match the set of methods we're using here. Pretty random.