@Crypt-iQ 2021-06-24
Finished reviewing, code looks good.  Have some more nits after comments are addressed.  I think `DataToSign` could also be bench-marked, but that doesn't have to be in this PR since it's already large.  Going to run the fuzz tests against this PR and see if anything breaks (I don't expect it to).  This PR does prep for buffer pool usage which is awesome!