@yyforyongyu 2021-07-02
> The diff reads well to me, and I've started to run this on one of my larger testnet nodes that has lndmon enabled so I can track metrics such as the heap size, total allocated/freed, etc. Will report back with my findings.

Just curious about how the heap allocs prior to this PR look like? Do we happen to have that kind of data?