@Roasbeef 2021-07-22
> Just curious about how the heap allocs prior to this PR look like? Do we happen to have that kind of data?

I mainly have some historical data from my own person `lndmon` instance...