@Crypt-iQ 2021-08-17
post-merge update: I ran the fuzz tests for lnwire and this pull does not break anything